// src/prisma/contract.d.ts
export type CodecTypes = PgTypes;
// ...
export type DiskOutput = {
  readonly count: CodecTypes['pg/float8@1']['output'] | null;
  readonly sizeGb: CodecTypes['pg/float8@1']['output'] | null;
  readonly type: CodecTypes['pg/text@1']['output'];
};
export type DiskInput = {
  readonly count: CodecTypes['pg/float8@1']['input'] | null;
  readonly sizeGb: CodecTypes['pg/float8@1']['input'] | null;
  readonly type: CodecTypes['pg/text@1']['input'];
};
export type SwapOutput = {
  readonly sizeGb: CodecTypes['pg/float8@1']['output'] | null;
  readonly kind: CodecTypes['pg/text@1']['output'];
};
// ...
export namespace Models {
  export type public_Build = {
    id: CodecTypes['pg/text@1']['output'];
    createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    romName: CodecTypes['pg/text@1']['output'];
    romVersion: CodecTypes['pg/text@1']['output'];
    androidVersion: CodecTypes['pg/float8@1']['output'] | null;
    hostType: CodecTypes['pg/text@1']['output'];
    hostingProvider: CodecTypes['pg/text@1']['output'];
    monthlyPriceUsd: CodecTypes['pg/float8@1']['output'] | null;
    networkSpeed: CodecTypes['pg/float8@1']['output'] | null;
    networkSpeedUnit: CodecTypes['pg/text@1']['output'];
    cpuModel: CodecTypes['pg/text@1']['output'];
    cpuCores: CodecTypes['pg/float8@1']['output'] | null;
    cpuThreads: CodecTypes['pg/float8@1']['output'] | null;
    memoryGb: CodecTypes['pg/float8@1']['output'] | null;
    memoryType: CodecTypes['pg/text@1']['output'];
    memorySpeedMhz: CodecTypes['pg/float8@1']['output'] | null;
    disks: ReadonlyArray<DiskOutput>;
    swaps: ReadonlyArray<SwapOutput>;
    buildMinutes: CodecTypes['pg/float8@1']['output'];
    dirtyBuildMinutes: CodecTypes['pg/float8@1']['output'] | null;
    kernelName: CodecTypes['pg/text@1']['output'];
    kernelVersion: CodecTypes['pg/text@1']['output'];
    kernelBuildMinutes: CodecTypes['pg/float8@1']['output'] | null;
    kernelDirtyBuildMinutes: CodecTypes['pg/float8@1']['output'] | null;
    notes: CodecTypes['pg/text@1']['output'];
    readonly [RelationKeys]?: never;
  };
}